Development

+++ src/ExpenseTracker.API/Controllers/PursesController.cs
@@ -5,8 +5,11 @@ using System.Threading.Tasks;
 using AutoMapper;
 using ExpenseTracker.API.Dtos.Purses;
 using ExpenseTracker.API.Repositories.Interfaces;
+using ExpenseTracker.Domain.Auth;
 using ExpenseTracker.Domain.Purses;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ExpenseTracker.API.Controllers
@@ -17,25 +20,41 @@ namespace ExpenseTracker.API.Controllers
     {
         private readonly IPurseRepository _repository;
         private readonly IMapper _mapper;
+        private UserManager<User> _userManager;
+        IAuthorizationService _authorizationService;
+        IPersonRepository _personRepository;
 
-        public PursesController(IPurseRepository repository, IMapper mapper)
+        public PursesController(IPurseRepository repository, IMapper mapper,
+            UserManager<User> userManager, IAuthorizationService authorizationService,
+            IPersonRepository personRepository)
         {
             _repository = repository;
             _mapper = mapper;
+            _userManager = userManager;
+            _authorizationService = authorizationService;
+            _personRepository = personRepository;
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetPurse(int id)
         {
             var purse = await _repository.Get(id);
-            if (purse != null)
+            if(purse == null)
             {
-                var purseDto = _mapper.Map<PurseDto>(purse);
-                return Ok(purseDto);
+                return NotFound();
             }
-            return NotFound();
+
+            var AR = await _authorizationService.AuthorizeAsync(HttpContext.User, purse, "Permission");
+            if (!AR.Succeeded)
+            {
+                return Unauthorized();
+            }
+
+            var purseDto = _mapper.Map<PurseDto>(purse);
+            return Ok(purseDto);
         }
 
+        [Authorize(Roles = "admin")]
         [HttpGet]
         public async Task<IActionResult> GetAllPurses()
         {
@@ -45,21 +64,42 @@ namespace ExpenseTracker.API.Controllers
             return Ok(pursesDto);
         }
 
-        [HttpGet("user/{id}")]
-        public async Task<IActionResult> GetUserPurses(int id)
+        [HttpGet("person/{id}")]
+        public async Task<IActionResult> GetPersonPurses(int id)
         {
             var purses = await _repository.Where(p => p.PersonId == id);
+
+            var AR = await _authorizationService.AuthorizeAsync(HttpContext.User, purses, "Permission");
+            if (!AR.Succeeded)
+            {
+                return Unauthorized();
+            }
+
             List<PurseDto> pursesDto = new List<PurseDto>();
             purses.ToList().ForEach(purse => pursesDto.Add(_mapper.Map<PurseDto>(purse)));
             return Ok(pursesDto);
         }
 
-        [HttpPost]
-        public async Task<IActionResult> CreatePurse([FromBody] PurseForUpdateDto purseForUpdateDto)
+        [HttpPost("person/{id}")]
+        public async Task<IActionResult> CreatePurse([FromBody] PurseForCreateDto purseForCreateDto, int id)
         {
-            var purse = PurseFactory.CreateEmptyPurse(purseForUpdateDto.CurrencyCode);
+            var person = await _personRepository.Get(id);
+            if(person == null)
+            {
+                return NotFound();
+            }
 
-            _mapper.Map(purseForUpdateDto, purse);
+            var AR = await _authorizationService.AuthorizeAsync(HttpContext.User, person, "Permission");
+            if (!AR.Succeeded)
+            {
+                return Unauthorized();
+            }
+
+            var user = await GetUserAsync();
+            var purse = PurseFactory.CreateEmptyPurse(purseForCreateDto.CurrencyCode);
+
+            _mapper.Map(purseForCreateDto, purse);
+            purse.OwnerId = user.Id;
 
             await _repository.Add(purse);
             await _repository.SaveChangesAsync();
@@ -69,13 +109,19 @@ namespace ExpenseTracker.API.Controllers
             return CreatedAtAction(nameof(GetPurse), new { id = purseDto.Id }, purseDto);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePurse(int id, PurseForUpdateDto purseForUpdateDto)
+        [HttpPut]
+        public async Task<IActionResult> UpdatePurse(PurseForUpdateDto purseForUpdateDto)
         {
-            var purse = await _repository.Get(id);
-            if (purse == null)
+            var purse = await _repository.Get(purseForUpdateDto.Id);
+            if(purse == null)
             {
                 return NotFound();
+            }
+
+            var AR = await _authorizationService.AuthorizeAsync(HttpContext.User, purse, "Permission");
+            if (!AR.Succeeded)
+            {
+                return Unauthorized();
             }
 
             _mapper.Map(purseForUpdateDto, purse);
@@ -94,10 +140,22 @@ namespace ExpenseTracker.API.Controllers
                 return NotFound();
             }
 
+            var AR = await _authorizationService.AuthorizeAsync(HttpContext.User, purse, "Permission");
+            if (!AR.Succeeded)
+            {
+                return Unauthorized();
+            }
+
             _repository.Remove(purse);
             await _repository.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private async Task<User> GetUserAsync()
+        {
+            var id = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "UserId").Value;
+            return await _userManager.FindByIdAsync(id);
         }
     }
 }
